Improved ensure_git_bash Better bash detection Minimized git install now requires input Moved literal_input function (Had to so I could confirm the minimized git install with it)
